requirements.txt
streamlit
pdfplumber
jinja2
python-dotenv
google-genai
requests
beautifulsoup4
